feat: run MACE model conversion in train/compress phase
