+++ arcann_training/common/utils.py
@@ -16,7 +16,7 @@ catch_errors_decorator(func: Callable[..., Any]) -> Callable[..., Any]
     Decorator to wrap a function and catch exceptions during execution.
 convert_seconds_to_hh_mm_ss(seconds: float) -> str
     Converts a time duration in seconds to the format HH:MM:SS.
-natural_sort_key(s: str) -> List[Union[int, str]]
+natural_sort_key(s: str) -> list[int | str]]
     Provides a natural sorting key for alphanumeric strings.
 """
 
@@ -24,7 +24,7 @@ natural_sort_key(s: str) -> List[Union[int, str]]
 import logging
 import re
 from functools import wraps
-from typing import Any, Callable, List, Union
+from typing import Any, Callable
 
 
 # Unittested
@@ -92,7 +92,7 @@ def convert_seconds_to_hh_mm_ss(seconds: float) -> str:
 
 # TODO: Add tests for this function
 @catch_errors_decorator
-def natural_sort_key(s: str) -> List[Union[int, str]]:
+def natural_sort_key(s: str) -> list[int | str]:
     """
     Provides a natural sorting key for alphanumeric strings.
 
@@ -103,7 +103,7 @@ def natural_sort_key(s: str) -> List[Union[int, str]]:
 
     Returns
     -------
-    List[Union[int, str]]
+    list[int | str]
         The sorting key.
 
     Raises
